Added a base62 encoder/decoder JSP tag and fixed a bug with long numbers.

+++ src/main/java/com/sangupta/jerry/encoder/Base62Encoder.java
@@ -59,9 +59,9 @@ public class Base62Encoder {
 		
 		int remainder;
 		do {
-			remainder = (int) number % 62;
+			remainder = (int) (number % 62l);
 			builder.append(elements[remainder]);
-			number = (int) number / 62;
+			number = number / 62;
 		} while(number > 0);
 		
 		return builder.toString();
